# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectX foundation API",
    "version": "1.0.0",
    "description": "ProjectX versioned API. People roster reads require a server-verified identity and organization scope."
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Process liveness",
        "responses": {
          "200": {
            "description": "Healthy process",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "ok"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/people/accounts/options": {
      "get": {
        "operationId": "getPeopleAddUserOptions",
        "summary": "Read actor-assignable Add User options",
        "description": "PROJECTX IMPLEMENTATION DECISION. Requires trusted organization user.manage; role and permission subsets reflect current effective rights. Each venue requires explicit current venue.manage. No browser-selected identity or organization; no SevenRooms semantics inferred.",
        "responses": {
          "200": { "description": "Authorized options, including empty groups", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PeopleAddUserOptions" } } } },
          "400": { "description": "Invalid query", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "401": { "description": "No trusted identity", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "403": { "description": "Missing current user.manage", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } },
          "500": { "description": "Internal failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/api/v1/people/accounts": {
      "post": {
        "operationId": "createPeopleAccount",
        "summary": "Create an organization-scoped pending People account",
        "description": "ProjectX implementation decision. Requires a server-trusted identity and organization-scoped user.manage. Creates User, OrganizationMembership, explicit authorized role/venue grants and a separate pending provisioning record atomically. The unverified email belongs only to the organization-scoped provisioning record; no OIDC identity, invitation delivery or MFA enrollment is performed.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._:-]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePeopleAccountRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pending ProjectX account provisioning created, or the identical idempotent result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePeopleAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No trusted identity or scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing current authority or a requested organization, venue, role or grant exceeds actor scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Organization-scoped pending email conflict or conflicting idempotency-key reuse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listPeopleAccounts",
        "summary": "Read authorized People roster",
        "description": "ProjectX implementation. Organization-scoped user.read is required; a server-trusted identity and organization are supplied outside the query. Only active memberships and organization role grants contribute. Venue-only grants do not authorize this roster. Empty authorized result is 200 with items []; pagination and UUID ordering are ProjectX decisions, not reference behavior.",
        "parameters": [
          {
            "name": "accessLevel",
            "in": "query",
            "required": false,
            "description": "Exact ProjectX organization role display name; filters only the authorized roster.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120,
              "pattern": "\\S"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Exclusive User ID cursor in ascending UUID order.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, default 25 and maximum 100.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized roster or empty collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleAccountsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No trusted identity or scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing current user.read organization capability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PeopleAccountSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "jobTitle",
          "emailNotificationsEnabled",
          "accessLevels"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Derived from persisted first and last name; null when both absent."
          },
          "jobTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "emailNotificationsEnabled": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "null means unconfigured; false explicitly disabled; true explicitly enabled."
          },
          "accessLevels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sorted, distinct names of active ProjectX organization role grants. No primary-role precedence or venue-role inference."
          }
        }
      },
      "PeopleAccountsResponse": {
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeopleAccountSummary"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Last returned User ID if another page exists; null otherwise."
          }
        }
      },
      "PeoplePermissionId": {
        "type": "string",
        "enum": ["session.read.self", "user.read.self", "user.read", "user.manage", "organization.manage", "venue.read", "venue.manage", "reservation.read", "reservation.create", "reservation.update", "reservation.cancel", "client.read", "client.create", "client.update", "client.history.read.organization", "floorplan.read", "floorplan.manage", "availability.read", "availability.manage", "import.manage", "client.export", "audit.read"]
      },
      "PeopleRoleOption": {
        "type": "object", "additionalProperties": false, "required": ["id", "name"],
        "properties": { "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" } }
      },
      "PeoplePermissionOption": {
        "type": "object", "additionalProperties": false, "required": ["id", "description", "scope"],
        "properties": { "id": { "$ref": "#/components/schemas/PeoplePermissionId" }, "description": { "type": "string" }, "scope": { "type": "string", "enum": ["SELF", "ORGANIZATION", "VENUE"] } }
      },
      "PeopleVenueOption": {
        "type": "object", "additionalProperties": false, "required": ["id", "name", "roles", "permissions"],
        "properties": {
          "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" },
          "roles": { "type": "array", "items": { "$ref": "#/components/schemas/PeopleRoleOption" } },
          "permissions": { "type": "array", "items": { "$ref": "#/components/schemas/PeoplePermissionOption" } }
        }
      },
      "PeopleAddUserOptions": {
        "type": "object", "additionalProperties": false, "required": ["organizationRoles", "organizationPermissions", "venues"],
        "properties": {
          "organizationRoles": { "type": "array", "items": { "$ref": "#/components/schemas/PeopleRoleOption" } },
          "organizationPermissions": { "type": "array", "items": { "$ref": "#/components/schemas/PeoplePermissionOption" } },
          "venues": { "type": "array", "items": { "$ref": "#/components/schemas/PeopleVenueOption" } }
        }
      },
      "CreatePeopleAccountVenueRequest": {
        "type": "object",
        "required": [
          "venueId",
          "roleIds"
        ],
        "additionalProperties": false,
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "permissionIds": {
            "type": "array", "maxItems": 22, "uniqueItems": true,
            "description": "Optional additive direct VENUE grants, checked against current actor rights at this explicit venue.",
            "items": { "$ref": "#/components/schemas/PeoplePermissionId" }
          },
          "roleIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "CreatePeopleAccountRequest": {
        "type": "object",
        "required": [
          "email",
          "firstName",
          "lastName",
          "jobTitle",
          "emailNotificationsEnabled",
          "suspended",
          "organizationRoleIds",
          "venues"
        ],
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "description": "Unverified organization-scoped provisioning email; never an OIDC identity key."
          },
          "firstName": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 120
          },
          "lastName": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 120
          },
          "jobTitle": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 120
          },
          "emailNotificationsEnabled": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "null means not established; false and true are explicit preferences."
          },
          "suspended": {
            "type": "boolean",
            "description": "ProjectX maps true to User.disabled_at at creation."
          },
          "organizationRoleIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "organizationPermissionIds": {
            "type": "array", "maxItems": 22, "uniqueItems": true,
            "description": "Optional additive direct ORGANIZATION/SELF grants owned by the new membership; no shared-role mutation.",
            "items": { "$ref": "#/components/schemas/PeoplePermissionId" }
          },
          "venues": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/CreatePeopleAccountVenueRequest"
            }
          }
        }
      },
      "CreatePeopleAccountResponse": {
        "type": "object",
        "required": [
          "provisioningId",
          "userId",
          "membershipId",
          "status"
        ],
        "additionalProperties": false,
        "properties": {
          "provisioningId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "const": "PENDING"
          }
        }
      }
    }
  }
}